+++ libraries/base/src/Data/Bifunctor.hs
@@ -133,39 +133,39 @@ class (forall a. Functor (p a)) => Bifunctor p where
     second = bimap id
 
 
--- | Class laws for tuples hold only up to laziness. Both
--- 'first' 'id' and 'second' 'id' are lazier than 'id' (and 'fmap' 'id'):
+-- | Tuple instances have the same laziness as for 'Functor'. Both
+-- 'first' 'id' and 'second' 'id' have the same laziness as 'id' (and 'fmap' 'id'):
 --
--- >>> first id (undefined :: (Int, Word)) `seq` ()
--- ()
--- >>> second id (undefined :: (Int, Word)) `seq` ()
--- ()
+-- >>> first id (errorWithoutStackTrace "error!" :: (Int, Word)) `seq` ()
+-- *** Exception: error!
+-- >>> second id (errorWithoutStackTrace "error!" :: (Int, Word)) `seq` ()
+-- *** Exception: error!
 -- >>> id (errorWithoutStackTrace "error!" :: (Int, Word)) `seq` ()
 -- *** Exception: error!
 --
 -- @since 4.8.0.0
 instance Bifunctor (,) where
-    bimap f g ~(a, b) = (f a, g b)
+    bimap f g (a, b) = (f a, g b)
 
 -- | @since 4.8.0.0
 instance Bifunctor ((,,) x1) where
-    bimap f g ~(x1, a, b) = (x1, f a, g b)
+    bimap f g (x1, a, b) = (x1, f a, g b)
 
 -- | @since 4.8.0.0
 instance Bifunctor ((,,,) x1 x2) where
-    bimap f g ~(x1, x2, a, b) = (x1, x2, f a, g b)
+    bimap f g (x1, x2, a, b) = (x1, x2, f a, g b)
 
 -- | @since 4.8.0.0
 instance Bifunctor ((,,,,) x1 x2 x3) where
-    bimap f g ~(x1, x2, x3, a, b) = (x1, x2, x3, f a, g b)
+    bimap f g (x1, x2, x3, a, b) = (x1, x2, x3, f a, g b)
 
 -- | @since 4.8.0.0
 instance Bifunctor ((,,,,,) x1 x2 x3 x4) where
-    bimap f g ~(x1, x2, x3, x4, a, b) = (x1, x2, x3, x4, f a, g b)
+    bimap f g (x1, x2, x3, x4, a, b) = (x1, x2, x3, x4, f a, g b)
 
 -- | @since 4.8.0.0
 instance Bifunctor ((,,,,,,) x1 x2 x3 x4 x5) where
-    bimap f g ~(x1, x2, x3, x4, x5, a, b) = (x1, x2, x3, x4, x5, f a, g b)
+    bimap f g (x1, x2, x3, x4, x5, a, b) = (x1, x2, x3, x4, x5, f a, g b)
 
 
 -- | @since 4.8.0.0